requirements-dev.txt
# Development dependencies for rail-django-graphql library
-r requirements.txt

# Testing
pytest>=8.0.0
pytest-django>=4.0.0
pytest-cov>=5.0.0
factory-boy>=3.3.0
coverage>=7.0.0

# Code quality
black>=23.0.0
isort>=5.0.0
flake8>=6.0.0
mypy>=1.0.0
djlint>=1.35.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=2.0.0
myst-parser>=2.0.0